refactor(termstructures): hoist shared sync_extrapolation helper
